token post requst works

+++ api/views.py
@@ -6,13 +6,27 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.authtoken.views import ObtainAuthToken
+from rest_framework.authtoken.models import Token
 
-from django.shortcuts import redirect
 
+class CustomAuthToken(ObtainAuthToken):
+    """[summary]
 
-def redirect_view(request):
-    response = redirect('/api-token-auth/')
-    return response
+    Args:
+        ObtainAuthToken ([type]): [description]
+    """
+
+    def post(self, request, *args, **kwargs):
+        serializer = self.serializer_class(data=request.data,
+                                           context={'request': request})
+        serializer.is_valid(raise_exception=True)
+        user = serializer.validated_data['user']
+        token, created = Token.objects.get_or_create(user=user)
+        return Response({
+            'token': token.key,
+            'user_id': user.pk,
+        })
 
 
 class MessageListView(APIView):
@@ -33,6 +47,14 @@ class MessageListView(APIView):
         return Response(serializer.data)
 
     def post(self, request):
+        """[summary]
+
+        Args:
+            request ([type]): [description]
+
+        Returns:
+            [type]: [description]
+        """
         data = request.data
         serializer = MessageSerializer(data=data)
         if serializer.is_valid() and len(data) > 0 and data["sender"] == request.user.id:
@@ -44,29 +66,25 @@ class MessageListView(APIView):
 class MessageDetailView(APIView):
     permission_classes = [IsAuthenticated]
 
-    # TODO fix try catch
     def get_object(self, request, message_id):
         try:
             return Message.objects.get(Q(id=message_id) & (Q(sender_id=request.user.id) | Q(receiver_id=request.user.id)))
         except Message.DoesNotExist:
-            raise Message.DoesNotExist
-
-    def get(self, request, id):
-        try:
-            message = self.get_object(request, id)
-        except Message.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
-        message.read()
+    def get(self, request, id):
+        message = self.get_object(request, id)
+        if(type(message) != Message):
+            return message
 
+        message.read()
         serializer = MessageSerializer(message)
         return Response(serializer.data)
 
     def put(self, request, id):
-        try:
-            message = self.get_object(request, id)
-        except Message.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+        message = self.get_object(request, id)
+        if(type(message) != Message):
+            return message
 
         serializer = MessageSerializer(message, data=request.data)
         if serializer.is_valid():
@@ -75,10 +93,10 @@ class MessageDetailView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
-        try:
-            message = self.get_object(request, id)
-        except Message.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+        message = self.get_object(request, id)
+        if(type(message) != Message):
+            return message
+
         message.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -97,7 +115,7 @@ class UserMessages(APIView):
         if request.method == 'GET':
             if len(request.body) > 0 and "isRead" in request.body.decode("utf-8"):
                 isReadFlag = JSONParser().parse(request)["isRead"]
-            if isReadFlag != None:
+            if isReadFlag is not None:
                 usersMessages = usersMessages.filter(Q(isRead=isReadFlag))
             serializer = MessageSerializer(usersMessages, many=True)
             return Response(serializer.data)
